zb,zm,zp,zx: Make use of new bus name types
This is to ensure that these string types are checked for conformance.
We should do that anyway but the D-Bus specification also mandates
implementations to verify verify everything to and from the bus.

A similar type for interface strings.

While generic impelmentations and `TryFrom` impelmentations ensure that
impact to users is minimum with this change, unfortunately they do now
have to specify type when passing a `None` value. This howeover is a
well-known pain-point in the Rust world and nothing specific to zbus.

Our currently solution is close to what will come in the future with `!`
types: https://github.com/rust-lang/rust/issues/35121 .